Move IntSym definition into symbolize module
This change moves the definition of the IntSym type into the symbolize
module. We already defined the Sym in there, so it makes little sense to
have a "more internal" type be defined at a higher level.
