+++ src/lib.rs
@@ -64,8 +64,6 @@ use std::fmt::Result as FmtResult;
 use std::num::NonZeroU32;
 use std::result;
 
-use resolver::IntSym;
-use resolver::SrcLang;
 use resolver::SymResolver;
